+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed full text search not using index.
+- Remove custom emojis that are no longer used in display name or bio.
 
 ## [5.6.0] - 2026-06-26